% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/norm_zi_switch.R
\name{alt_EM_ctrl}
\alias{alt_EM_ctrl}
\title{Runs the EM algorithm for zero inflated model}
\usage{
alt_EM_ctrl(y, t, maxit = 500, loglik_tol = 1e-07, verbose = FALSE)
}
\arguments{
\item{y}{Gene expression vector}

\item{t}{Pseudotime vector}

\item{verbose}{Print out EM progress}
}
\value{
List with parameters, latent values and log likelihood
}
\description{
Runs the EM algorithm for zero inflated model
}

