% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/common.R
\name{calc_mu}
\alias{calc_mu}
\title{Calculate the mean vector given parameters and pseudotimes}
\usage{
calc_mu(params, t)
}
\arguments{
\item{params}{Vector of length 4 with entries L = 2mu_0, k, t0, sig_sq}

\item{t}{Vector of pseudotimes}
}
\value{
Mean sigmoidal vector
}
\description{
This function (common to all models) calculates the sigmoidal mean vector
given the parameters and factor of pseudotimes
}

