% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/switch.R
\name{diffExprSCE}
\alias{diffExprSCE}
\title{Differential expression test across pseudotime}
\usage{
diffExprSCE(sce, use = "expr", dist = c("nn", "normal"))
}
\description{
Differential expression test across pseudotime
}

