% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/switchde.R
\docType{data}
\name{ex_pseudotime}
\alias{ex_pseudotime}
\title{Synthetic gene pseudotimes}
\format{An object of class \code{array} of length 100.}
\usage{
ex_pseudotime
}
\value{
A vector of length 100
}
\description{
A vector with example pseudotimes for the synthetic 
gene expression data in \code{example_gex}
}
\keyword{datasets}
