% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/switchde.R
\docType{data}
\name{example_gex}
\alias{example_gex}
\title{Synthetic gene expression matrix}
\format{An object of class \code{matrix} with 4 rows and 100 columns.}
\usage{
example_gex
}
\value{
A 4 by 100 matrix
}
\description{
A matrix containing some synthetic gene expression data for 
100 cells and 4 genes
}
\keyword{datasets}

