% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/switchde.R
\name{example_sigmoid}
\alias{example_sigmoid}
\title{Example sigmoid plot}
\usage{
example_sigmoid()
}
\value{
An object of class \code{ggplot}
}
\description{
Plot an example sigmoid function. For demonstration and documentation.
}
\examples{
example_sigmoid()
}
