% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sctools.R, R/switch.R
\name{fitModel}
\alias{fitModel}
\title{Returns a set of fitted sigmoidal pseudotime models}
\usage{
fitModel(x, t = NULL, dist = c("nb", "normal"), sigmoidal = TRUE)

fitModel(x, t = NULL, dist = c("nb", "normal"), sigmoidal = TRUE)
}
\description{
Returns a set of fitted sigmoidal pseudotime models

Fit pseudotime model
}

