% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/switch.R
\name{log_neg_bin_likelihood}
\alias{log_neg_bin_likelihood}
\title{Log-likelihood of a negative binomial function with count vector k,
mean vector mu and constant dispersion parameter r}
\usage{
log_neg_bin_likelihood(k, mu, r)
}
\description{
Log-likelihood of a negative binomial function with count vector k,
mean vector mu and constant dispersion parameter r
}

