% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/common.R
\name{log_norm_likelihood}
\alias{log_norm_likelihood}
\title{Log-likelihood of iid measurements from a normal distribution}
\usage{
log_norm_likelihood(x, mu, sig_sq)
}
\arguments{
\item{x}{A vector of measurements}

\item{mu}{A scalar mean}

\item{sig_sq}{a scalar variance}
}
\value{
The log likelihood x given mu, sig_sq
}
\description{
Log-likelihood of iid measurements from a normal distribution
}

