% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/switch.R
\name{nb_alt_obj_func}
\alias{nb_alt_obj_func}
\title{Objective function given differential expression}
\usage{
nb_alt_obj_func(params, x, t)
}
\description{
Objective function given differential expression
}

