% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normal_switch.R
\name{norm_diff_expr_test}
\alias{norm_diff_expr_test}
\title{Sigmoidal differential expression test}
\usage{
norm_diff_expr_test(x, t, ...)
}
\arguments{
\item{x}{Gene expression vector}

\item{t}{Pseudotime vector}

\item{...}{Additional arguments passed to \code{norm_fit_models}}
}
\value{
A vector of length 5 with entries:
\itemize{
\item P-value
\item MLE estimate for L = 2 mu_0 parameter
\item MLE estimate for k
\item MLE estimate for t_0
\item MLE estimate for sigma^2
}
}
\description{
Returns the p-value and model for sigmoidal differential expression. 
Non-zero-inflated and Gaussian likelihood
}

