% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/normal_switch.R
\name{norm_fit_alt_model}
\alias{norm_fit_alt_model}
\title{Fit the sigmoidal expression model for some expression vector
x, pseudotime t, and control parameters to be passed to optim
This function returns NA if optim fails to converge.}
\usage{
norm_fit_alt_model(x, t, control = list(maxit = 1e+05))
}
\description{
Fit the sigmoidal expression model for some expression vector
x, pseudotime t, and control parameters to be passed to optim
This function returns NA if optim fails to converge.
}

