% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normal_switch.R
\name{norm_fit_null_model}
\alias{norm_fit_null_model}
\title{Fit null model}
\usage{
norm_fit_null_model(x)
}
\arguments{
\item{x}{Gene expression vector}
}
\value{
A list of length three, with entries:
\itemize{
\item par MLE parameter estimates
\item log_lik The MLE log-likelihood
\item value The value returned by \code{optim} (equivalent to negative log-likelihood)
}
}
\description{
Fits the null model (simply by taking the mean and 
variance of the expression vector x) and returns a list
with the parameters as first entry and negative log likelihood
as second.
}

