% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normal_switch.R
\name{norm_plot_model}
\alias{norm_plot_model}
\title{Plot a sigmoidal pseudotime model}
\usage{
norm_plot_model(alt_model, x, t)
}
\arguments{
\item{alt_model}{The model returned by norm_fit_alt_model}

\item{x}{Gene expression vector x}

\item{t}{Pseudotime vector t}
}
\value{
A ggplot object plotting both gene expression and predicted
gene expression (e.g. the mean) over pseudotime
}
\description{
Plot a sigmoidal pseudotime model
}

