% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normal_switch.R
\name{norm_Q_grad}
\alias{norm_Q_grad}
\title{Compute the gradient of the log-likelihood}
\usage{
norm_Q_grad(params, x, t)
}
\arguments{
\item{params}{A vector with three entries: (1) Parameter \eqn{L = 2 \mu_0}, (2) \eqn{k} 
and (3) \code{t_0}}

\item{x}{Gene expression vector}

\item{t}{Pseudotime vector}
}
\value{
The function gradient with respect to \eqn{(L, k, t_0)}.
}
\description{
Computations the gradient of the log-likelihood of the model
for params, expression vector x and pseudotime t
}

