% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/norm_zi_switch.R
\name{norm_zi_lrtest}
\alias{norm_zi_lrtest}
\title{Likelihood ratio test for zero-inflated sigmoidal differential expression}
\usage{
norm_zi_lrtest(x, t, models)
}
\arguments{
\item{x}{Gene expression vector}

\item{t}{Pseudotime vector}

\item{models}{List of length two, with the first entry corresponding
to the sigmoidal model and the latter to the null model. The model should
be of the form of those returned by norm_fit_alt_model and norm_fit_null_model}
}
\value{
A P-value given by the likelihood ratio test
}
\description{
Likelihood ratio test for zero-inflated sigmoidal differential expression
}

