% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/norm_zi_switch.R
\name{norm_zi_plot_model}
\alias{norm_zi_plot_model}
\title{Plot zero inflated sigmoidal model}
\usage{
norm_zi_plot_model(params, x, t)
}
\arguments{
\item{params}{Parameters returned by \code{EMCtrl}}

\item{x}{Gene expression vector}

\item{t}{Pseudotime vector}
}
\description{
Plot zero inflated sigmoidal model
}

