% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm_zi_switch.R
\name{norm_zi_Q}
\alias{norm_zi_Q}
\title{Return the Q function}
\usage{
norm_zi_Q(params, x, t, is_zero)
}
\arguments{
\item{params}{parameter vector}

\item{x}{Gene expression vector}

\item{t}{Pseudotime vector}

\item{is_zero}{Logical vector indicating which elements of x were originally zero}
}
\value{
The estimate of the Q function.
}
\description{
Return the Q function
}

