% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm_zi_switch.R
\name{norm_zi_Q_grad}
\alias{norm_zi_Q_grad}
\title{Return the gradient of Q = E[l_c]}
\usage{
norm_zi_Q_grad(params, x, t, is_zero)
}
\arguments{
\item{params}{parameter vector}

\item{x}{Gene expression vector}

\item{t}{Pseudotime vector}

\item{is_zero}{Logical vector indicating which elements of x were originally zero}
}
\description{
Return the gradient of Q = E[l_c]
}

