% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/switchde.R
\name{sigmoid}
\alias{sigmoid}
\title{Calculate the mean vector given parameters and pseudotimes (mu0 formulation)}
\usage{
sigmoid(pst, params)
}
\arguments{
\item{pst}{Vector of pseudotimes}

\item{params}{Vector of length 3 with entries mu_0, k, t0}
}
\value{
Mean sigmoidal vector
}
\description{
This function (common to all models) calculates the sigmoidal mean vector
given the parameters and factor of pseudotimes
}
\keyword{internal}
