% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/generate.R
\name{simulateCells}
\alias{simulateCells}
\title{Simulate single-cell RNA-seq data}
\usage{
simulateCells(ngenes, hypers, pst = NULL, ncells = NULL, dropout = TRUE,
  rate = NULL, base = 10)
}
\description{
Simulate single-cell RNA-seq data
}

