% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/switchde.R
\name{switchde}
\alias{switchde}
\title{Switch-like model fitting and differential expression test}
\usage{
switchde(object, pseudotime = NULL, zero_inflated = FALSE, ...)
}
\arguments{
\item{object}{Gene expression data that is either
\itemize{
 \item A vector of length number of cells for a single gene
 \item A matrix of dimension number of genes x number of cells
 \item An object of class \code{SCESet} from package scater
 }}

\item{pseudotime}{A pseudotime vector with a pseudotime corresponding to 
every cell. Can be \code{NULL} if object is of class \code{SCESet} and 
\code{pData(sce)$pseudotime} is defined.}

\item{zero_inflated}{Logical. Should zero inflation be implemented? Default  \code{FALSE}}

\item{...}{Additional arguments to be passed to expectation maximisation algorithm
if zero-inflation is enabled:
\itemize{
 \item maxit Maximum number of iterations for EM. Default 500
 \item loglik_tol Change in log-likelihood tolerance. Default 1e-7 
 \item verbose Should progress of EM optimisation be printed? Default \code{FALSE}
 }}
}
\value{
A matrix where each column corresponds to a gene, the first row is
the p-value for that gene and subsequent rows are model parameters.
}
\description{
Fit sigmoidal differential expression models to gene expression across pseudotime.
Parameter estimates are returned along with a p-value for switch-like differential
expression over a null model (constant expression).
}
\examples{
data(synth_gex)
data(ex_pseudotime)
sde <- switchde(synth_gex, ex_pseudotime)
}

