% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/switchde.R
\docType{data}
\name{synth_gex}
\alias{synth_gex}
\title{Synthetic gene expression matrix}
\format{An object of class \code{matrix} with 12 rows and 100 columns.}
\usage{
synth_gex
}
\value{
A 12 by 100 matrix
}
\description{
A matrix containing some synthetic gene expression data for 
100 cells and 12 genes
}
\keyword{datasets}
