% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/switchde.R
\name{testDE}
\alias{testDE}
\title{Backwards compatibility.}
\usage{
testDE(...)
}
\arguments{
\item{...}{Arguments passed to \code{switchde}.}
}
\description{
This function as been replaced by \code{switchde}.
}

