% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/switchde.R
\name{fitModel}
\alias{fitModel}
\title{Fit sigmoidal pseudotime model for set of genes}
\usage{
fitModel(object, pseudotime = NULL, zero_inflated = FALSE, ...)
}
\value{
A matrix where columns are genes and rows are model parameters (zero inflation will
have an extra lambda parameter compared to no zero-inflation)
}
\description{
@param object Gene expression data that is either
 \itemize{
 \item A vector of length number of cells for a single gene
 \item A matrix of dimension number of genes x number of cells
 \item An object of class \code{SCESet} from package scater
 }
 @param pseudotime A pseudotime vector with a pseudotime corresponding to 
 every cell. Can be \code{NULL} if object is of class \code{SCESet} and 
 \code{pData(sce)$pseudotime} is defined.
 @param zero_inflated Logical. Should zero inflation be implemented? Default  \code{FALSE}
 @param ... Additional arguments to be passed to expectation maximisation algorithm
 if zero-inflation is enabled:
 \itemize{
 \item maxit Maximum number of iterations for EM. Default 500
 \item loglik_tol Change in log-likelihood tolerance. Default 1e-7 
 \item verbose Should progress of EM optimisation be printed? Default \code{FALSE}
 }
}

