% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R, R/switch.R
\name{norm_alt_obj_func}
\alias{norm_alt_obj_func}
\title{Negative log likelihood of the sigmoidal differential expression
function given some expression vector x}
\usage{
norm_alt_obj_func(params, x, t)

norm_alt_obj_func(params, x, t)
}
\description{
Negative log likelihood of the sigmoidal differential expression
function given some expression vector x

Objective function given differential expression
}

