% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/normal_switch.R
\name{norm_fit_models}
\alias{norm_fit_models}
\title{Fits both sigmoidal and null models, returning them as a list}
\usage{
norm_fit_models(x, t)
}
\description{
Fits both sigmoidal and null models, returning them as a list
}

