% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normal_switch.R
\name{norm_Q_grad}
\alias{norm_Q_grad}
\title{Computations the gradient of the log-likelihood of the model
for params, expression vector x and pseudotime t}
\usage{
norm_Q_grad(params, x, t)
}
\description{
Computations the gradient of the log-likelihood of the model
for params, expression vector x and pseudotime t
}

