% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/switchde.R
\name{sanitise_inputs}
\alias{sanitise_inputs}
\title{Sanitise inputs}
\usage{
sanitise_inputs(object, pseudotime, lower_threshold, zero_inflated, sce_assay)
}
\arguments{
\item{object}{The object passed at the entry point (either a SCESet or gene
expression matrix)}

\item{pseudotime}{A pseudotime vector}

\item{lower_threshold}{The minimum threshold below which to set expression to zero to avoid
numerical issues. Default is 0.01}

\item{zero_inflated}{Logical. Should zero inflation be implemented? Default  \code{FALSE}}

\item{sce_assay}{The assay from the \code{SingleCellExperiment} to be used
as expression, defaulting to "exprs"}
}
\value{
A list with two entries: a gene expression matrix \code{X}
and a pseudotime vector \code{pst}.
}
\description{
Sanitise inputs
}
\keyword{internal}
